DeepLeGATo++/requirements_colab.txt
# DeepLeGATo++ Requirements for Google Colab

# Core ML
torch>=2.1.0
pytorch-lightning>=2.1.0
timm>=0.9.12

# Normalizing Flows for NPE
nflows>=0.14
zuko>=1.0.0

# Astronomy
astropy>=6.0.0
photutils>=1.10.0
reproject>=0.11.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Experiment tracking
wandb>=0.16.0

# Colab utilities
gdown>=4.7.0
ipywidgets>=8.0.0

# Web demo
gradio>=4.0.0

# Utilities
pyyaml>=6.0.0
tqdm>=4.66.0
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
einops>=0.7.0
